+++ tests/Database/Functional/Driver/Postgres/Query/InsertQueryTest.php
@@ -128,18 +128,18 @@ class InsertQueryTest extends CommonClass
 
     public function testInsertMicroseconds(): void
     {
-        $schema = $this->schema(table: 'with_microseconds', driverConfig: ['datetimeWithMicroseconds' => true]);
+        $schema = $this->schema(table: 'with_microseconds', driverConfig: ['withDatetimeMicroseconds' => true]);
         $schema->primary('id');
         $schema->datetime('datetime', 6);
         $schema->save();
 
         $expected = new \DateTimeImmutable();
 
-        $id = $this->db(driverConfig: ['datetimeWithMicroseconds' => true])->insert('with_microseconds')->values([
+        $id = $this->db(driverConfig: ['withDatetimeMicroseconds' => true])->insert('with_microseconds')->values([
             'datetime' => $expected,
         ])->run();
 
-        $result = $this->db(driverConfig: ['datetimeWithMicroseconds' => true])->select('datetime')
+        $result = $this->db(driverConfig: ['withDatetimeMicroseconds' => true])->select('datetime')
             ->from('with_microseconds')
             ->where('id', $id)
             ->run()