+++ CHANGELOG.md
@@ -5,7 +5,7 @@ v2.3.0 (in progress)
 - Add supporting for array values in the `IN` and `NOT IN` operators by @roxblnfk (#69, #70, #71)
 - Add `PdoInterface` as possible return type of the `Driver::getPDO()` method by @roxblnfk (#76)
 - Add array parameter `options` to all driver configs to pass additional driver options.
-  To store a DateTime with microseconds, you can pass the `datetimeWithMicroseconds` with value **true**
+  To store a DateTime with microseconds, you can pass the `withDatetimeMicroseconds` with value **true**
   in the new options parameter by @msmakouz (#86)
 - Add ability to define `size` in the `datetime` column by @msmakouz (#86)